src/app/app.component.spec.ts
import { CommonModule } from '@angular/common';
import { TestBed } from '@angular/core/testing';
import { BrowserModule } from '@angular/platform-browser';
import { AppInstructionComponent } from './app-instruction/app-instruction.component';
import { AppComponent } from './app.component';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';

describe('AppComponent', () => {
  let httpMock: HttpTestingController;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [
        AppComponent,
        AppInstructionComponent
      ],
      imports: [
        HttpClientTestingModule, CommonModule, BrowserModule
      ],
    }).compileComponents();
  });

  it('should create the app', () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.componentInstance;
    expect(app).toBeTruthy();
  });

  it('should display task instructions', () => {
    const fixture = TestBed.createComponent(AppComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement;
    expect(compiled.querySelector('h1').textContent).toContain('Angular Single Responsibility Principle exercise');
  });

  it('should fetch user data from server', () => {
    const fixture = TestBed.createComponent(AppComponent);
    fixture.detectChanges();
    httpMock = TestBed.inject(HttpTestingController);

    const mockReq = httpMock.expectOne((req) => {
      return req.method === 'GET';
    });

    mockReq.flush({});
    httpMock.verify();
  });
});
